Add support for new alert box types
Current settings in puter only supported the warning alert box
type and there was no ability pass in custom parameters.

Added 4 new alert box types: info, success, question, error.
And added styling specific to each of these alert types.
The ability now existing to specific an alert type when
calling the alert system. The default alert type if none
is passed in is 'warning.'

Added ability for custom html string parameters to be passed
into the alert system. If a custom html string is passed,
it overrides the alert message.

Resolves: Issue #3